+++ src/chat/server.clj
@@ -3,6 +3,7 @@
   (:require [buddy.hashers :as hashers]
             [clojure.core.async :as a]
             [clojure.java.io :as io]
+            [java-time.api :as jt]
             [clojure.string :as str]
             [hato.client :as hc]
             [muuntaja.middleware :as mw]
@@ -17,11 +18,11 @@
             [ring.websocket.keepalive :as wska]
             [taoensso.telemere :as t]
             ;;
-            [chat.xtdb :as xt]
-            ))
+            [chat.xtdb :as xt]))
+
 (def debug? (System/getenv "MX3_DEV"))
 
-(def ^:private version "v0.12.116")
+(def ^:private version "v0.13.136")
 
 (def ^:private l22
   (if debug?
@@ -108,21 +109,39 @@
               {:as :json :timeout 1000})
       :body))
 
-;; no effect.
-;; (defn- wrap-debug [handler]
-;;   (fn [request]
-;;     (t/log! :debug (:body request))
-;;     (handler request)))
+(defn- load-data [{{:keys [n]} :path-params}]
+  (xt/q '{:find [author message timestamp]
+          :keys [author message timestamp]
+          :in [t0]
+          :where [[e :author author]
+                  [e :message message]
+                  [e :timestamp timestamp]
+                  [(<= t0 timestamp)]]}
+        (jt/minus (jt/local-date-time) (jt/minutes (Long/parseLong n)))))
+
+;; (defn- load-data [{{:keys [n]} :path-params :as request}]
+;;   (def *r* request)
+;;   (xt/q '{:find [(pull eid [*])]
+;;           ;; :keys [author message timestamp]
+;;           :in [t0]
+;;           :where [[eid :timestamp timestamp]
+;;                   [(<= t0 timestamp)]]}
+;;         (jt/minus (jt/local-date-time) (jt/minutes (Long/parseLong n)))))
+
+(comment
+  (load-data {:path-params {:n "10"}})
+  :rcf)
 
 (defn make-app-handler []
   (rr/ring-handler
-   (rr/router [["/chat" {:middleware [;; wrap-debug
-                                      [wst/wrap-websocket-transit]
+   (rr/router [["/chat" {:middleware [[wst/wrap-websocket-transit]
                                       [wska/wrap-websocket-keepalive]]}
                 ["" (make-chat-handler)]]
                ["/api" {:middleware [[def/wrap-defaults def/api-defaults]
                                      mw/wrap-format
                                      mw/wrap-params]}
+                ["/load/:n" (fn [req]
+                              (resp/response (load-data req)))]
                 ["/user-random" {:get (fn [_]
                                         (resp/response (user-random nil)))}]]
                ["" {:middleware [[def/wrap-defaults def/site-defaults]]}